Node selection and dataflow viz

+++ main.py
@@ -1,5 +1,6 @@
 import tree
 from tree_viz import TreeVisualization
+from tree_viz import tree_nodes
 from dataflow_viz import DataflowVisualization
 from atrib_viz import AttributesVisualization
 import setup
@@ -14,6 +15,7 @@ if __name__ == "__main__":
     kdd = tree.get_dataset('file')
 
     curr_scene_id = 1
+    node_id = None
 
     while True:
 
@@ -23,10 +25,23 @@ if __name__ == "__main__":
             curr_scene_id = 1
             viz = TreeVisualization(my_tree, kdd)
 
+            # pick a node using mouse click - then press 2 to show dataflow viz.
+            def get_node():
+                global node_id
+                pick = curr_scene.mouse.pick
+                if pick is not None:
+
+                    for node in tree_nodes:
+                        if node[0].pos == pick.pos:
+                            node_id = node[5]
+                            pick.emissive = True
+
+            curr_scene.bind("mousedown", get_node)
+
         if curr_scene_id == 2:
             curr_scene = setup.setup_scene("Tok dát\n")
             curr_scene_id = 2
-            viz = DataflowVisualization(my_tree, curr_scene)
+            viz = DataflowVisualization(my_tree, curr_scene, node_id)
 
         if curr_scene_id == 3:
             curr_scene = setup.setup_scene("Vplyv a frekvencia atribútov\n")
@@ -38,7 +53,7 @@ if __name__ == "__main__":
             k = keysdown()
             setup.move_camera(k, curr_scene)
             new_scene = setup.switch_scene(k)
-            if new_scene != curr_scene_id and new_scene != -1:
+            if new_scene != curr_scene_id and new_scene != -1 and node_id is not None:
                 curr_scene_id = new_scene
                 curr_scene.title = ''
                 curr_scene.delete()
